# docs/swagger.yaml
basePath: /
definitions:
...
  model.CaseUpdate:
    properties:
      arrivedDate:
        type: string
      caseDetail:
        type: string
      caseDuration:
        type: integer
      caseId:
        type: string
      caseLat:
        type: string
      caseLon:
        type: string
      caseSTypeId:
        type: string
      caseTypeId:
        type: string
      caseVersion:
        type: string
      caselocAddr:
        type: string
      caselocAddrDecs:
        type: string
      closedDate:
        type: string
      commandedDate:
        type: string
      countryId:
        type: string
      createdDate:
        type: string
      deviceId:
        type: string
      distId:
        type: string
      extReceive:
        type: string
      phoneNo:
        type: string
      phoneNoHide:
        type: boolean
      priority:
        type: integer
      provId:
        type: string
      receivedDate:
        type: string
      referCaseId:
        type: string
      resDetail:
        type: string
      resId:
        type: string
      scheduleDate:
        type: string
      scheduleFlag:
        type: boolean
      source:
        type: string
      startedDate:
        type: string
      statusId:
        type: string
      userarrive:
        type: string
      userclose:
        type: string
      usercommand:
        type: string
      usercreate:
        type: string
      userreceive:
        type: string
      wfId:
        type: string
    type: object
...
  model.CommandInsert:
    properties:
      active:
        description: Active status
        type: boolean
      deptId:
        description: Department ID
        type: string
      en:
        description: English name
        type: string
      th:
        description: Thai name
        type: string
    type: object
  model.CommandUpdate:
    properties:
      active:
        description: Active status
        type: boolean
      deptId:
        description: Department ID
        type: string
      en:
        description: English name
        type: string
      th:
        description: Thai name
        type: string
    type: object
...
  model.Response:
    properties:
      data: {}
      desc:
        type: string
      msg:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    email: support@somewhere.com
    name: API Support
    url: http://somewhere.com/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://somewhere.com/
  title: CMS API
  version: "1.0"
paths:
...
  /api/v1/case/{id}:
    delete:
      consumes:
      - application/json
      operationId: Delete Case
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Case
      tags:
      - Cases
    get:
      consumes:
      - application/json
      operationId: Case By Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Cases By Id
      tags:
      - Cases
    patch:
      consumes:
      - application/json
      operationId: Update Case
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CaseUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Case
      tags:
      - Cases
...
  /api/v1/commands:
    get:
      consumes:
      - application/json
      operationId: Get Commands
      parameters:
      - default: 0
        description: start
        in: query
        name: start
        type: integer
      - default: 10
        description: length
        in: query
        name: length
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Commands
      tags:
      - Organization
  /api/v1/commands/{id}:
    delete:
      consumes:
      - application/json
      operationId: Delete Commands
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Commands
      tags:
      - Organization
    get:
      consumes:
      - application/json
      operationId: Get Commands by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Commands by id
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      operationId: Update Commands
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CommandUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Commands
      tags:
      - Organization
  /api/v1/commands/add:
    post:
      consumes:
      - application/json
      operationId: Create Commands
      parameters:
      - description: Create Data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CommandInsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Request successful
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Commands
      tags:
      - Organization
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"